requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
numpy>=1.24.0,<2.0
tensorflow>=2.15.0
